fix(glue): arm bulk-OUT on open; complete pico2 example; fix host test
The USBTMC glue re-armed the bulk-OUT endpoint after every receive,
IN-completion and clear, but never armed the *first* read: that is left
to tud_usbtmc_open_cb, which the glue did not define. Consumers had to
remember to call tud_usbtmc_start_bus_read() in their own open_cb; when
they didn't, enumeration succeeded but every host write timed out with
-110 (ETIMEDOUT). The earlier "-110 fix" commit only addressed the IN
path, so this OUT-path -110 remained.

Changes:
- glue: define tud_usbtmc_open_cb() -> tud_usbtmc_start_bus_read() so the
  middleware owns the first arm; consumers must no longer define open_cb.
- tests: add regression coverage that open_cb arms the bus read.
- tests: fix a signed-char comparison in test_usbscpi that failed on
  platforms where char is signed (x86); the data check was already
  covered by the adjacent memcmp.
- examples/pico2: add the strong-symbol USBTMC callbacks TinyUSB requires
  (get_capabilities/abort/clear/msgBulkOut_start) and a tusb_config.h, so
  the example actually links. It previously could not build.
- README: document that the glue owns open/bus-read, and list the
  callbacks the application must still provide.

Host tests (test_usbscpi, test_glue_tinyusb) pass. The glue change is the
same one verified on ESP32-S3 hardware (relocated from the app into the
glue). The pico2 example was not compile-tested (no Pico SDK available).

+++ tests/test_usbscpi.c
@@ -156,7 +156,9 @@ static void test_param_parsing_and_arbitrary_block(void) {
     assert(f.tx[2] == '1');
     assert(f.tx[3] == '0');
     assert(memcmp(&f.tx[4], block_data_buf, 10) == 0);
-    assert(f.tx[4 + 10 - 1] == block_data_buf[9]);
+    /* f.tx is a char[]; cast through uint8_t so the high-bit payload byte
+     * (0xF6) compares equal on platforms where char is signed (x86). */
+    assert((uint8_t)f.tx[4 + 10 - 1] == block_data_buf[9]);
     assert(f.tx[4 + 10] == '\n');
 }
